Switch to PSR-4 for tests
