+++ tests/Fixtures/FacebookTestProviderStub.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace Tests\Fixtures;
+
+use Mockery as m;
+use Laravel\Socialite\Two\FacebookProvider;
+
+class FacebookTestProviderStub extends FacebookProvider
+{
+    public $http;
+
+    protected function getUserByToken($token)
+    {
+        return ['id' => 'foo'];
+    }
+
+    /**
+     * Get a fresh instance of the Guzzle HTTP client.
+     *
+     * @return \GuzzleHttp\Client
+     */
+    protected function getHttpClient()
+    {
+        if ($this->http) {
+            return $this->http;
+        }
+
+        return $this->http = m::mock('StdClass');
+    }
+}
