+++ tests/Fixtures/OAuthOneTestProviderStub.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Fixtures;
+
+use Laravel\Socialite\One\AbstractProvider;
+
+class OAuthOneTestProviderStub extends AbstractProvider
+{
+    //
+}
